+++ angel_container/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.0-alpha.10
 * Added `Container.registerLazySingleton<T>`.
+* Added named singleton support.
 
 # 1.0.0-alpha.9
 * Added `Container.has<T>`.